feat(slider): add fluent ui range slider
