+++ package.json
@@ -42,6 +42,7 @@
     "@babel/preset-env": "7.16.11",
     "@babel/preset-react": "7.16.7",
     "@babel/preset-typescript": "7.16.7",
+    "@fluentui/react": "8.61.2",
     "@mdx-js/react": "1.6.22",
     "@stylelint/postcss-css-in-js": "0.37.2",
     "@types/enzyme": "3.10.11",
