+++ packages/docs/src/mdx/coreComponents/Slider.mdx
@@ -10,6 +10,7 @@ import styled from 'styled-components'
 import {
   Slider,
   SliderField,
+  RangeSliderDemo,
   SpaceBlock,
 } from 'practical-react-components-core'
 
@@ -89,6 +90,8 @@ export const DemoComponent = () => {
 ## Demo
 
 <DemoComponent />
+<SpaceBlock variant={16} />
+<RangeSliderDemo />
 
 ## Basic usage
